Resolved three conflicts from master's 'Unify types with Geth' PR (#11937),
which migrated gas/nonce types from long/UInt256 to ulong codebase-wide:

GethLikeTxTrace.cs:
  - Master changed Gas long → ulong and kept StoragesByDepth.
  - Our branch had removed StoragesByDepth (no longer used — storage is
    tracked per-entry via GethTxTraceEntry.Storage since PR #12056).
  - Resolution: take ulong Gas, keep StoragesByDepth removed.

VirtualMachineTests.cs:
  - Master added (ulong) casts in MCOPY gas assertions because
    GasCostOf constants are now ulong (ulong * int won't compile).
  - Our branch changed Memory from string[] to UInt256[], so memory
    assertions compare UInt256 values, not hex strings.
  - Resolution: take master's (ulong) casts for gas; keep our UInt256
    comparisons for memory.

DebugRpcModuleTests.TraceBlock.cs:
  - Master reformatted struct-log JSON (removed extra alignment spaces,
    added "error": null, added "storage": {} on every step).
  - Our branch emits "error" only when non-null and "storage" only at
    SLOAD/SSTORE steps with the actual cumulative snapshot.
  - Resolution: keep our branch's semantics; adopt master's spacing fix.

Also fixed GethLikeTxTraceJsonLinesConverter.WriteHexLong(long) → ulong
to match the auto-merged GethTxTraceEntry.Gas/GasCost ulong types.

+++ src/Nethermind/Nethermind.Evm.Test/VirtualMachineTests.cs
@@ -482,7 +482,7 @@ public class VirtualMachineTests : VirtualMachineTestsBase
             .Done;
         GethLikeTxTrace traces = Execute(new GethLikeTxMemoryTracer(Build.A.Transaction.TestObject, GethTraceOptions.Default), code, MainnetSpecProvider.CancunActivation).BuildResult();
 
-        Assert.That(traces.Entries[^2].GasCost, Is.EqualTo(GasCostOf.VeryLow + GasCostOf.VeryLow * ((data.Length + 31) / 32) + GasCostOf.Memory * 0), "gas");
+        Assert.That(traces.Entries[^2].GasCost, Is.EqualTo(GasCostOf.VeryLow + GasCostOf.VeryLow * (ulong)((data.Length + 31) / 32) + GasCostOf.Memory * 0UL), "gas");
     }
 
     [Test]
@@ -505,7 +505,7 @@ public class VirtualMachineTests : VirtualMachineTestsBase
 
         using (Assert.EnterMultipleScope())
         {
-            Assert.That(traces.Entries[^2].GasCost, Is.EqualTo(GasCostOf.VeryLow + GasCostOf.VeryLow * ((data.Length + 31) / 32) + GasCostOf.Memory * 1), "gas");
+            Assert.That(traces.Entries[^2].GasCost, Is.EqualTo(GasCostOf.VeryLow + GasCostOf.VeryLow * (ulong)((data.Length + 31) / 32) + GasCostOf.Memory * 1UL), "gas");
             Assert.That(origin, Is.EqualTo(copied));
         }
     }
@@ -531,7 +531,7 @@ public class VirtualMachineTests : VirtualMachineTestsBase
 
         using (Assert.EnterMultipleScope())
         {
-            Assert.That(traces.Entries[^2].GasCost, Is.EqualTo(GasCostOf.VeryLow + GasCostOf.VeryLow * (SLICE_SIZE + 31) / 32), "gas");
+            Assert.That(traces.Entries[^2].GasCost, Is.EqualTo(GasCostOf.VeryLow + GasCostOf.VeryLow * (ulong)(SLICE_SIZE + 31) / 32), "gas");
             Assert.That(result, Is.EqualTo(new UInt256(Bytes.FromHexString("0x0101020304050607080000000000000000000000000000000000000000000000"), isBigEndian: true)), "memory state");
         }
     }
@@ -553,7 +553,7 @@ public class VirtualMachineTests : VirtualMachineTestsBase
 
         using (Assert.EnterMultipleScope())
         {
-            Assert.That(traces.Entries[^2].GasCost, Is.EqualTo(GasCostOf.VeryLow + GasCostOf.VeryLow * ((data.Length + 31) / 32)), "gas");
+            Assert.That(traces.Entries[^2].GasCost, Is.EqualTo(GasCostOf.VeryLow + GasCostOf.VeryLow * (ulong)((data.Length + 31) / 32)), "gas");
             Assert.That(traces.Entries.Last().Memory.Count, Is.EqualTo(1));
         }
     }
@@ -591,7 +591,7 @@ public class VirtualMachineTests : VirtualMachineTestsBase
 
         using (Assert.EnterMultipleScope())
         {
-            Assert.That(traces.Entries[^2].GasCost, Is.EqualTo(GasCostOf.VeryLow + GasCostOf.VeryLow * (SLICE_SIZE + 31) / 32), "gas");
+            Assert.That(traces.Entries[^2].GasCost, Is.EqualTo(GasCostOf.VeryLow + GasCostOf.VeryLow * (ulong)(SLICE_SIZE + 31) / 32), "gas");
             Assert.That(result, Is.EqualTo(new UInt256(Bytes.FromHexString("0x0102030405060708080000000000000000000000000000000000000000000000"), isBigEndian: true)), "memory state");
         }
     }
